Add logic To review View and ReviewTableViewcell
add function setupCell to set reviewCell his value
add implementation of TableView delegate and dataSource
