+++ e-store/e-store/ScenesScreen/Register/RegisterAccount/RegisterAccountScreenViewController.swift
@@ -30,6 +30,7 @@ class RegisterAccountScreenViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
+       
     }
     
 }
